@@ -19,7 +19,7 @@ public class PassengerDto {
     // 7ajtou b7aja o5ra ya3ml request :p
     //*
 
-    public PassengerDto fromEntity(Passenger passenger){
+    public static PassengerDto fromEntity(Passenger passenger){
         if (passenger == null)
             throw new InvalidResource(null,"INTERNAL ERROR","Can't map null entity");
         return PassengerDto.builder()
