@@ -5,8 +5,9 @@ import com.team.fithniti.demo.model.RideRequest;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.querydsl.QuerydslPredicateExecutor;
 
 import java.util.UUID;
 
-public interface RideRepo extends JpaRepository<Ride,Long> {
+public interface RideRepo extends JpaRepository<Ride,Long>, QuerydslPredicateExecutor<Ride> {
 }
